Add Remove method to RootFS
Call Remove from the integration test and also check that the directory that the root fs was using can be deleted.

Add a syscall integration test.

Note that Generate may have a minor side-effect on error.

Use %s instead of %q for printing gerrors to improve formatting.

Use a more reasonable default filemode for created directories